Rename TForce `RatesPackageOptions`
This renames `RatesPackageOptions` to `PackageOptions` for consistency
with the other services. The renamed class is aliased for backwards
compatibility, but also marked as deprecated.
