+++ lib/friendly_shipping/services/tforce_freight.rb
@@ -7,6 +7,7 @@ require 'friendly_shipping/services/tforce_freight/shipping_methods'
 require 'friendly_shipping/services/tforce_freight/shipment_document'
 require 'friendly_shipping/services/tforce_freight/shipment_information'
 require 'friendly_shipping/services/tforce_freight/rates_options'
+require 'friendly_shipping/services/tforce_freight/package_options'
 require 'friendly_shipping/services/tforce_freight/rates_package_options'
 require 'friendly_shipping/services/tforce_freight/item_options'
 require 'friendly_shipping/services/tforce_freight/rates_item_options'
