+++ spec/friendly_shipping/services/tforce_freight/generate_pickup_request_hash_spec.rb
@@ -69,7 +69,7 @@ RSpec.describe FriendlyShipping::Services::TForceFreight::GeneratePickupRequestH
 
     let(:package_options) do
       [
-        FriendlyShipping::Services::TForceFreight::RatesPackageOptions.new(
+        FriendlyShipping::Services::TForceFreight::PackageOptions.new(
           package_id: package_one.id,
           item_options: item_one_options
         )
